simplify down to one conftest.py
