+++ tests/integration/__init__.py
@@ -1,0 +1,1 @@
+pytest_plugins = "bokeh.testing.plugins.integration_tests"
